# MediaLibrary/MediaLibrary.Client/Api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "MediaLibrary.API",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Query/tracks-in-album": {
      "get": {
        "tags": [
          "Query"
        ],
        "summary": "Возвращает список треков в альбоме",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "Идентификатор альбома",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Track"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Track"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Track"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Query/albums-info": {
      "get": {
        "tags": [
          "Query"
        ],
        "summary": "Возвращает альбомы и количество треков в каждом",
        "parameters": [
          {
            "name": "year",
            "in": "query",
            "description": "Год альбома",
            "schema": {
              "maximum": 2024,
              "minimum": 1900,
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AlbumInfoDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AlbumInfoDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AlbumInfoDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AlbumInfoDto": {
        "required": [
          "actorId",
          "date",
          "name",
          "trackCount"
        ],
        "type": "object",
        "properties": {
          "actorId": {
            "type": "integer",
            "description": "Идентификатор исполнителя",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "description": "Название Альбома",
            "nullable": true
          },
          "date": {
            "type": "string",
            "description": "Дата релиза",
            "format": "date"
          },
          "trackCount": {
            "type": "integer",
            "description": "Количество треков в альбоме",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "Dto для альбома с количеством треков"
      },
...
      "Track": {
        "required": [
          "albumId",
          "id",
          "name",
          "number",
          "time"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "albumId": {
            "type": "integer",
            "format": "int32"
          },
          "time": {
            "type": "string",
            "format": "date-span"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}